Fixing long communication

+++ src/VCluster_unit_tests.hpp
@@ -360,7 +360,7 @@ BOOST_AUTO_TEST_CASE( VCluster_use_sendrecv)
 		{
 			global_step = j;
 			// Processor step
-			size_t ps = n_proc / (8 + 1);
+			long int ps = n_proc / (8 + 1);
 
 			// send message
 			openfpm::vector<openfpm::vector<unsigned char>> message;
@@ -411,9 +411,9 @@ BOOST_AUTO_TEST_CASE( VCluster_use_sendrecv)
 #endif
 
 			// Check the message
-			for (size_t i = 0 ; i < 8  && i < n_proc ; i++)
+			for (long int i = 0 ; i < 8  && i < n_proc ; i++)
 			{
-				long int p_id = (- (i+1) * ps + vcl.getProcessUnitID()) % n_proc;
+				long int p_id = (- (i+1) * ps + (long int)vcl.getProcessUnitID()) % n_proc;
 				if (p_id < 0)
 					p_id += n_proc;
 				else
